Correctif espion

+++ VikingRaider/Assets/Scripts/Actions.cs
@@ -21,7 +21,7 @@ public class Actions : MonoBehaviour {
         {
             town.garni_known.is_known = true;
             town.garni_known.turn_known = time.currentTurn;
-            float percent = town.garnison.number / 100;
+            float percent = (float) town.garnison.number / 100f;
             int success = UnityEngine.Random.Range(100 - quality, 100 + quality);
             town.garni_known.value_known = (int)Math.Floor((float) success * percent);
             //Urgal : espionnage de la garnison
@@ -63,7 +63,11 @@ public class Actions : MonoBehaviour {
                 //Urgal : delete espion
                 UIManager.DrawSpyResult(0, ResultType.DEADSPY, spy);
             }
-            else { getinfo(town, spy, time); }
+            else
+            {
+                getinfo(town, spy, time);
+                //todo faire la distinction de cas
+            }
         }
         else { getinfo(town, spy, time); }
